--- src/Responsive.Core/src/Abstractions/IResponsiveFactory.cs
@@ -1,9 +1,0 @@
-// Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
-// The Apache v2. See License.txt in the project root for license information.
-
-namespace Wangkanai.Responsive
-{
-    public interface IResponsiveFactory
-    {
-    }
-}
