+++ src/Responsive.Core/src/Builders/ResponsiveApplicationBuilderExtensions.cs
@@ -61,7 +61,7 @@ namespace Wangkanai.Responsive
         [Obsolete("This will be refactor to add service of responsive in 2.0-beta-14. USe AddResponsive(options => {}) instead.")]
         public static IApplicationBuilder UseResponsive(
             this IApplicationBuilder app,
-            Action<IResponsiveOptionsBuilder> options)
+            Action<IResponsiveOptions> options)
         {
             if (app == null)
                 throw new UseResponsiveAppArgumentNullException(nameof(app));
